Automated text ui test
It sort of works but converting to UTF8 sucks. even with forcing intellij to use utf8 or whatever i found in google. the second run with the same exact code gives an error that says UTF8 is the wrong encoding type. (might come back later or smth to do this or not but it's sucking up so much of my time)